Remove support for swift versions below 5.9, and use a single package file Also removed all the code needed for compatibility with older versions
